backend/requirements.txt
fastapi==0.129.0
passlib==1.7.4
PyJWT==2.10.1
PyJWT==2.11.0
SQLAlchemy==2.0.46
strawberry==3.0
